pyproject.toml
[project]
name = "teaching-assistance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "logging>=0.4.9.6",
    "pika>=1.3.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "uuid>=1.30",
]
